changed the radix-4 to 16 bit

+++ matlab_implementation/encoder_algo/booth_encoder_radix_4.m
@@ -1,12 +1,12 @@
 function [encoded_array] = booth_encoder_radix_2(variable,bin_size)
-	variable = uint16(variable);
+	variable = uint32(variable);
 	encoded_array = zeros(1,bin_size/2);
 	
 	variable = bitshift(variable,1);
 	for index=0:2:bin_size-1
-		lb_0 = bitget(variable,bin_index(index),'uint16');
-		lb_1 = bitget(variable,bin_index(index+1),'uint16');
-		lb_2 = bitget(variable,bin_index(index+2),'uint16');
+		lb_0 = bitget(variable,bin_index(index),'uint32');
+		lb_1 = bitget(variable,bin_index(index+1),'uint32');
+		lb_2 = bitget(variable,bin_index(index+2),'uint32');
 		%booth_encoded= bitget(variable,0);
 		if (lb_2==0 && lb_1==0 && lb_0==1) || ( lb_2==0 && lb_1==1 && lb_0==0) 
 			encoded_array(bin_index(index/2)) = 1;
